answer trade limits on every edge instead of casting to a capability
